+++ NEWS.md
@@ -1,9 +1,10 @@
-# fhircrackr 2.2.1
+# fhircrackr 2.3.0
 
 ## Bug fixes
 - Errors when removing tags in `fhir_search()` (e.g. with `rm_tag = "div"`) now get caught and converted to a warning.
 
 ## New functions
+- `fhir_is_empty()` checks if a bundle or bundlelist is empty
 - `fhir_collapse()` collapses multiple entries that belong to the same higher level FHIR element and should be kept together, e.g. `name.given` or `address.line`
 
 - `fhir_melt_all()` divides all multiple entries into separate rows by repeatedly calling `fhir_melt()`.